refactor(lib): split supabase.ts god-class into per-domain db modules
The 1637-line PhraseExtractionService static class is broken into focused
modules under src/lib/db/ (shows, episodes, extractions, phrases,
extraction-jobs, stats, dedup) as plain async functions. Database schema
types move to src/types/database.ts and the client to supabase-client.ts.

src/lib/supabase.ts is now a thin barrel: re-exports the client + types and
assembles the PhraseExtractionService facade from the db modules, so all 26
existing import sites keep working unchanged. Internal this.x() calls are
rewired to explicit cross-module imports.

Verified: tsc clean, lint 0 errors, production build (76 pages incl. DB-backed
SSG) succeeds.
